+++ constants.go
@@ -1,12 +1,21 @@
+// Copyright (c) 2018, The TurtleCoin Developers
+//
+// Please see the included LICENSE file for more information.
+//
+
 package main
 
 const (
-	versionNest                 = "0.34"
+	versionNest                 = "0.35"
 	urlCryptoCompareTRTL        = "https://min-api.cryptocompare.com/data/price?fsym=TRTL&tsyms=USD"
-	defaultRemoteDaemonAddress  = "public.turtlenode.io"
-	defaultRemoteDaemonPort     = "11898"
 	logFileFilename             = "TurtleCoin-Nest.log"
 	urlBlockExplorer            = "https://blocks.turtle.link/"
 	dbFilename                  = "settings.db"
 	numberTransactionsToDisplay = 20
+
+	// default remote node
+	defaultRemoteDaemonAddress = "public.turtlenode.io"
+	defaultRemoteDaemonPort    = "11898"
+	defaultRemoteDaemonName    = "public.turtlenode.io"
+	defaultRemoteDaemonSSL     = false
 )