+++ platformSpecific_windows.go
@@ -1,5 +1,10 @@
 // +build windows
 
+// Copyright (c) 2018, The TurtleCoin Developers
+//
+// Please see the included LICENSE file for more information.
+//
+
 package main
 
 var isPlatformDarwin = false