+++ walletdmanager/walletdmanager.go
@@ -1,3 +1,8 @@
+// Copyright (c) 2018, The TurtleCoin Developers
+//
+// Please see the included LICENSE file for more information.
+//
+
 // Package walletdmanager handles the management of the wallet and the communication with the core wallet software
 package walletdmanager
 
@@ -122,7 +127,7 @@ func RequestAddress() (address string, err error) {
 // RequestListTransactions provides the list of transactions of current wallet
 func RequestListTransactions() (transfers []turtlecoinwalletdrpcgo.Transfer, err error) {
 
-	walletBlockCount, _, _, err := turtlecoinwalletdrpcgo.RequestStatus(rpcPassword)
+	walletBlockCount, _, _, _, err := turtlecoinwalletdrpcgo.RequestStatus(rpcPassword)
 	if err != nil {
 		log.Error("error getting block count: ", err)
 		return nil, err
@@ -263,7 +268,8 @@ func SaveWallet() (err error) {
 // walletPath is the full path to the wallet
 // walletPassword is the wallet password
 // useRemoteNode is true if remote node, false if local
-func StartWalletd(walletPath string, walletPassword string, useRemoteNode bool, daemonAddress string, daemonPort string) (err error) {
+// useCheckpoints is true if TurtleCoind should be run with "--load-checkpoints"
+func StartWalletd(walletPath string, walletPassword string, useRemoteNode bool, useCheckpoints bool, daemonAddress string, daemonPort string) (err error) {
 
 	if isWalletdRunning() {
 		errorMessage := "turtle-service is already running in the background.\nPlease close it via "
@@ -286,6 +292,8 @@ func StartWalletd(walletPath string, walletPassword string, useRemoteNode bool, 
 	pathToLogTurtleCoindAllSessions := logTurtleCoindAllSessionsFilename
 	pathToWalletd := "./" + walletdCommandName
 	pathToTurtleCoind := "./" + turtlecoindCommandName
+	checkpointsCSVFile := "checkpoints.csv"
+	pathToCheckpointsCSV := "./" + checkpointsCSVFile
 
 	WalletFilename = filepath.Base(walletPath)
 	pathToWallet := filepath.Clean(walletPath)
@@ -305,6 +313,7 @@ func StartWalletd(walletPath string, walletPassword string, useRemoteNode bool, 
 		pathToAppContents := filepath.Dir(pathToAppDirectory)
 		pathToWalletd = pathToAppContents + "/" + walletdCommandName
 		pathToTurtleCoind = pathToAppContents + "/" + turtlecoindCommandName
+		pathToCheckpointsCSV = pathToAppContents + "/" + checkpointsCSVFile
 
 		usr, err := user.Current()
 		if err != nil {
@@ -325,6 +334,7 @@ func StartWalletd(walletPath string, walletPassword string, useRemoteNode bool, 
 	} else if isPlatformLinux {
 		pathToWalletd = pathToAppDirectory + "/" + walletdCommandName
 		pathToTurtleCoind = pathToAppDirectory + "/" + turtlecoindCommandName
+		pathToCheckpointsCSV = pathToAppDirectory + "/" + checkpointsCSVFile
 		pathToLogWalletdCurrentSession = pathToAppDirectory + "/" + logWalletdCurrentSessionFilename
 		pathToLogWalletdAllSessions = pathToAppDirectory + "/" + logWalletdAllSessionsFilename
 		pathToLogTurtleCoindCurrentSession = pathToAppDirectory + "/" + logTurtleCoindCurrentSessionFilename
@@ -361,7 +371,11 @@ func StartWalletd(walletPath string, walletPassword string, useRemoteNode bool, 
 		}
 		defer turtleCoindCurrentSessionLogFile.Close()
 
-		cmdTurtleCoind = exec.Command(pathToTurtleCoind, "--log-file", pathToLogTurtleCoindCurrentSession)
+		if useCheckpoints {
+			cmdTurtleCoind = exec.Command(pathToTurtleCoind, "--load-checkpoints", pathToCheckpointsCSV, "--log-file", pathToLogTurtleCoindCurrentSession)
+		} else {
+			cmdTurtleCoind = exec.Command(pathToTurtleCoind, "--log-file", pathToLogTurtleCoindCurrentSession)
+		}
 		hideCmdWindowIfNeeded(cmdTurtleCoind)
 
 		turtleCoindAllSessionsLogFile, err := os.Create(pathToLogTurtleCoindAllSessions)
@@ -469,7 +483,7 @@ func StartWalletd(walletPath string, walletPassword string, useRemoteNode bool, 
 	}
 
 	// check rpc connection with walletd
-	_, _, _, err = turtlecoinwalletdrpcgo.RequestStatus(rpcPassword)
+	_, _, _, _, err = turtlecoinwalletdrpcgo.RequestStatus(rpcPassword)
 	if err != nil {
 		killWalletd()
 		return errors.New("error communicating with turtle-service via rpc")
@@ -764,33 +778,34 @@ func CreateWallet(walletFilename string, walletPassword string, walletPasswordCo
 }
 
 // RequestConnectionInfo provides the blockchain sync status and the number of connected peers
-func RequestConnectionInfo() (syncing string, blockCount int, knownBlockCount int, peerCount int, err error) {
+func RequestConnectionInfo() (syncing string, walletBlockCount int, knownBlockCount int, localDaemonBlockCount int, peerCount int, err error) {
 
-	blockCount, knownBlockCount, peerCount, err = turtlecoinwalletdrpcgo.RequestStatus(rpcPassword)
+	walletBlockCount, knownBlockCount, localDaemonBlockCount, peerCount, err = turtlecoinwalletdrpcgo.RequestStatus(rpcPassword)
 	if err != nil {
-		return "", 0, 0, 0, err
+		return "", 0, 0, 0, 0, err
 	}
 
-	stringWait := " (No transfers allowed until synced)"
+	stringWait := " (No transfers allowed)"
 	if knownBlockCount == 0 {
 		WalletdSynced = false
 		syncing = "Getting block count..." + stringWait
-	} else if blockCount < knownBlockCount-1 || blockCount > knownBlockCount+10 {
+		//} else if walletBlockCount < knownBlockCount-1 || walletBlockCount > knownBlockCount+10 || localDaemonBlockCount < knownBlockCount-1 || localDaemonBlockCount > knownBlockCount+10 {
+	} else if walletBlockCount < knownBlockCount-1 || walletBlockCount > knownBlockCount+10 {
 		// second condition handles cases when knownBlockCount is off and smaller than the blockCount
 		WalletdSynced = false
-		syncing = "Wallet syncing..." + stringWait
+		syncing = "Syncing..." + stringWait
 	} else {
 		WalletdSynced = true
-		syncing = "Wallet synced"
+		syncing = "Synced"
 	}
 
-	return syncing, blockCount, knownBlockCount, peerCount, nil
+	return syncing, walletBlockCount, knownBlockCount, localDaemonBlockCount, peerCount, nil
 }
 
 // RequestFeeinfo provides the additional fee requested by the remote node for each transaction
 func RequestFeeinfo() (nodeFee float64, err error) {
 
-	_, nodeFee, _, err = turtlecoinwalletdrpcgo.Feeinfo(rpcPassword)
+	_, nodeFee, _, err = turtlecoinwalletdrpcgo.GetFeeInfo(rpcPassword)
 	if err != nil {
 		return 0, err
 	}